added new file yamlVariable.go that contains logic for getting and parsing variables from flag,
