+++ README.md
@@ -47,7 +47,7 @@ Tip: | in YAML is for multiline text
 - [x] `--configdir` flag to let the user specify path to custom config folder. Fallback to default path if empty, also let user get the config from current working directory
 - [x] Remember custom config directory if specified AND `--remember` is true (by default false)
 - [x] Additional error handling for new features
-- [ ] YAML variables that can be set using a flag. $MY_VAR: Test, a: {MY_VAR}. Parse and replace {MY_VAR} with the actual value.
+- [x] YAML variables that can be set using a flag. $MY_VAR: Test, a: {MY_VAR}. Parse and replace {MY_VAR} with the actual value.
 
 ## Install
