+++ utils/scaffold.go
@@ -10,17 +10,21 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-var variableMatchPattern = regexp.MustCompile(`\$\w+`)
+// pattern to find all appearance of curly braces enclosed instances
+var variableMatchPattern = regexp.MustCompile(`\{[^{}]+\}`)
 
-func replaceVariable(name string, variableMap map[string]string) string {
+// replaceVariables takes a string content and a map containin set varibales with their values
+// the string that is returned has all occurences of a variable refrence replaced with the actual value that was set
+func replaceVariables(name string, variableMap map[string]string) string {
 	matches := variableMatchPattern.FindAllString(name, -1)
+	//check if there is at least a pattern match
 	if len(matches) <= 0 {
 		return name
 	}
-
+	//loop through collected variables set and replace any occurence
 	for k, v := range variableMap {
 		//construct regex pattern using our varibale key
-		pattern := fmt.Sprintf(`\$\b%s\b`, regexp.QuoteMeta(k))
+		pattern := fmt.Sprintf(`\{%s\}`, regexp.QuoteMeta(k))
 		// Compile the regular expression
 		regex := regexp.MustCompile(pattern)
 		name = regex.ReplaceAllString(name, v)
@@ -77,22 +81,20 @@ func Scaffold(name string, yamlpath string, setVariables map[string]string) {
 
 	// Scaffold the directory structure :: iterating over the map
 	for folder, files := range dirs {
-		//replace yaml variables if any
-		folder = replaceVariable(folder, setVariables)
 		// Create the folders and subdirectories if necessary
 		err = os.MkdirAll(folder, 0755)
 		helper.Fatal(fmt.Sprintf("Error creating folder %s: %v", folder, err), true, err)
 
 		// Create the files :: iterating over files from the map and getting specified content
 		for fileName, content := range files {
-			//replace yaml variables if any
-			fileName = replaceVariable(fileName, setVariables)
 			// Construct a file path for the file
 			filePath := filepath.Join(folder, fileName)
 			// Create the directories before creating the file
 			err = os.MkdirAll(filepath.Dir(filePath), 0755)
 			helper.Fatal(fmt.Sprintf("Error creating directories for %s: %v", filePath, err), true, err)
 
+			//replace all variables with values set
+			content = replaceVariables(content, setVariables)
 			// Create the files at filePath path and add specified content
 			err = os.WriteFile(filePath, []byte(content), 0644)
 			helper.Fatal(fmt.Sprintf("Failed to create file %s: %s", fileName, err), true, err)
